requirements.txt

Django==2.2
django-admin-rangefilter==0.5.3
django-allauth==0.39.1
django-appconf==1.0.3
django-babel==0.6.2
django-cache-machine==1.1.0
django-ckeditor==5.6.1
django-colorful==1.3
django-compressor==2.2
django-cors-headers==3.0.2
django-extensions==2.1.6
django-filter==2.1.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-jenkins==0.110.0
django-js-asset==1.2.2
django-maintenance-mode==0.14.0
django-microservices==0.1
django-mptt==0.9.1
django-multiselectfield==0.1.8
django-mysql==3.5.0
django-nose==0.1.3
django-redis-cache==2.0.0
django-rosetta==0.9.2
django-selenium==0.9.8
django-uuidfield==0.5.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
hiredis==1.0.0
hpack==3.0.0
html2text==2018.1.9
httplib2==0.12.1
hyper==0.7.0
hyperframe==3.2.0
idna==2.6
imagesize==1.1.0
importlib-metadata==0.8
inflection==0.5.0
invoke==1.2.0
ipaddress==1.0.22
ipdb==0.12
ipython==7.4.0
ipython-genutils==0.2.0
isort==4.3.16
itypes==1.2.0
jdcal==1.4
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
joblib==0.14.1
jsonfield==2.0.2
jsonpickle==1.1
lazy-object-proxy==1.3.1
lxml==4.3.2
mailsnake==1.6.4
Markdown==3.1
MarkupSafe==1.1.1
maxminddb==1.4.1
mccabe==0.6.1
microsofttranslator==0.8
mysqlclient==1.4.6
newrelic==4.16.0.116
nose==1.3.7
numpy==1.17.0
oauth2==1.9.0.post1
oauth2client==4.1.3
oauthlib==3.0.1
odfpy==1.4.0
olefile==0.46
opencv-python==4.2.0.32
opengraph==0.5
opengraph-py3==0.71
openpyxl==2.6.1
ordereddict==1.1
packaging==19.0
pandas==0.24.2
paramiko==2.4.2
parso==0.3.4
path.py==11.5.0
pefile==2018.8.8
pep8==1.7.1
pexpect==4.6.0
phone-iso3166==0.3.5
phonenumbers==8.10.14
pickleshare==0.7.5
Pillow==5.4.1
polib==1.1.0
profanity-check==1.0.3
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycparser==2.19
pyfcm==1.4.7
pygeoip==0.3.2
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.3.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.3.1
pysolr==3.8.1
pytesseract==0.2.6
python-datauri==0.2.8
python-dateutil==2.8.0
python-gflags==3.1.2
python-linkedin-v2==0.9.4
python-memcached==1.59
python-ptrace==0.9.3
python-redis-lock==3.3.1
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1
raven==6.10.0
rcssmin==1.0.6
redis==3.2.1
redis-structures==0.1.7
refreshbooks==2.0
reportlab==3.5.13
requests==2.21.0
requests-oauthlib==1.2.0
rjsmin==1.0.12
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.2.0
Savoir==1.0.6
scikit-learn==0.22.2.post1
scipy==1.4.1
selenium==3.141.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.2.1
soupsieve==1.9
Sphinx==1.8.5
sphinx-rtd-theme==0.4.3
sphinxcontrib-websupport==1.1.0
sqlparse==0.3.0
suds2==0.7.1
tablib==0.13.0
tlslite==0.4.9
tlslite-ng==0.7.5
topicaxis-opengraph==0.5
traitlets==4.3.2
twython==3.7.0
typed-ast==1.3.1
typing==3.6.6
tzlocal==1.5.1
unicodecsv==0.14.1
unittest-xml-reporting==2.5.1
uritemplate==3.0.0
urllib3==1.22
validators==0.15.0
vobject==0.9.6.1
wcwidth==0.1.7
webcolors==1.8.1
webpreview==1.6.0
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0
zipp==0.3.3
gunicorn